refactor: move queries files to queries folder
