+++ src/queries/byText.ts
@@ -4,8 +4,8 @@ import { matches, TextMatch } from '../matches';
 import type { NormalizerFn } from '../matches';
 import { makeQueries } from './makeQueries';
 import type { Queries } from './makeQueries';
-import { filterNodeByType } from './filterNodeByType';
-import { createLibraryNotSupportedError } from './errors';
+import { filterNodeByType } from '../helpers/filterNodeByType';
+import { createLibraryNotSupportedError } from '../helpers/errors';
 
 export type TextMatchOptions = {
   exact?: boolean;
